requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.8
langchain>=0.0.200
python-dotenv>=0.19.0
huggingface-hub>=0.16.4
transformers>=4.30.0
torch>=2.0.0 